Rename service module
