+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/service/get_service_class_decls.rs
